[WIP] Switch frames to be collectables.
Danger: highly explosive branch; very unlikley to build NQP, though
passes all but 2 of the NQP test suite. Pushed simply for backup.

The eventual goal is to manage frames that are not referenced at all
by other collectables in their life time using a traditional-style
call stack, and promote them to the heap as normal GC-able objects
upon escape. As a stepping stone, we'll allocate all frames on the
heap, to help uncover missing mark bugs and the like, as well as to
break the big change into steps between which stuff should entirely
work.